requirements.txt
fastapi==0.110.2
pydantic==2.7.1
sqlalchemy==2.0.29
pyjwt==2.8.0
bcrypt==4.1.2
uvicorn==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9